uploader/requirements.txt
aiodns==3.0.0
aiohttp[speedups]==3.7.4.post0
async-generator==1.10; python_version >= '3.5'
async-timeout==3.0.1; python_full_version >= '3.5.3'
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
biliup==0.1.8
brotlipy==0.7.0
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0; os_name == 'nt' and implementation_name != 'pypy'
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
charset-normalizer==2.0.7; python_version >= '3'
colorama==0.4.4; sys_platform == 'win32'
cryptography==35.0.0; python_version >= '3.6'
h11==0.12.0; python_version >= '3.6'
idna==3.3; python_version >= '3'
iso-639==0.4.5
iso3166==2.0.2; python_version >= '3.6'
iso8601==0.1.16
isodate==0.6.0
loguru==0.5.3
lxml==4.6.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
m3u8==0.9.0; python_version >= '3.5'
multidict==5.2.0; python_version >= '3.6'
mutagen==1.45.1; python_version >= '3.5' and python_version < '4'
outcome==1.1.0; python_version >= '3.6'
pillow==8.3.2; python_version >= '3.6'
psutil==5.8.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyasn1==0.4.8
pycares==4.1.2
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pycryptodome==3.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pycryptodomex==3.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pydantic==1.8.2
pyopenssl==21.0.0
pysocks==1.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.19.1
pyyaml==6.0; python_version >= '3.6'
requests==2.26.0
rsa==4.7.2
selenium==4.0.0; python_version ~= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0; python_version >= '3.5'
sortedcontainers==2.4.0
streamlink==2.4.0; python_version >= '3.6' and python_version < '4'
trio-websocket==0.9.2; python_version >= '3.5'
trio==0.19.0; python_version >= '3.6'
typing-extensions==3.10.0.2
urllib3[secure]==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
websocket-client==1.2.1; python_version >= '3.6'
websockets==10.0; python_version >= '3.7'
win32-setctime==1.0.3; sys_platform == 'win32'
wsproto==1.0.0; python_full_version >= '3.6.1'
yarl==1.7.0; python_version >= '3.6'
ykdl==1.7.2
youtube-dl==2021.6.6
yt-dlp==2021.10.22
